requirements.txt
# Web Framework
fastapi==0.103.2
uvicorn==0.22.0

# AWS SDK
boto3==1.24.17
botocore==1.27.18

# Data Processing
numpy==1.19.2
pandas==1.1.3
scipy==1.5.3
joblib==1.1.0

# Machine Learning
scikit-learn==0.23.2
imbalanced-learn==0.7.0
xgboost==1.5.2
statsmodels==0.12.2

# Visualization
matplotlib==3.3.4
seaborn==0.11.1

# SageMaker SDK
sagemaker==2.117.0

# Extra Utilities
requests==2.31.0
prometheus_client==0.16.0
# Web Framework
fastapi==0.103.2
uvicorn==0.22.0

# AWS SDK
boto3==1.24.17
botocore==1.27.18

# Data Processing
numpy==1.19.2
pandas==1.1.3
scipy==1.5.3
joblib==1.1.0

# Machine Learning
scikit-learn==0.23.2
imbalanced-learn==0.7.0
xgboost==1.5.2
statsmodels==0.12.2

# Visualization
matplotlib==3.3.4
seaborn==0.11.1

# SageMaker SDK
sagemaker==2.117.0

# Extra Utilities
requests==2.31.0
prometheus_client==0.16.0
python-dotenv==0.21.1